+++ CHANGES.rst
@@ -26,6 +26,7 @@ Bug fix:
  * Xcode11: Respect system dark mode setting (#2628).
  * Xcode11: Fix noisy notifications (#3316).
  * Xcode11: Temporary workaround for navigation bar bg color on emoji selection screen (#3271).
+ * Xcode11: Pass eventId when navigating to room from notification (#3321).
  
  Changes in 0.11.6 (2020-xx-xx)
  ===============================================